requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
alembic==1.7.6
aniso8601==9.0.1
asgiref==3.5.0
async-timeout==4.0.2
asyncio==3.4.3
asynctest==0.13.0
attrs==21.4.0
azure-cognitiveservices-vision-computervision==0.9.0
azure-common==1.1.28
azure-core==1.22.1
azure-storage-blob==12.9.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.3
cryptography==36.0.1
dnspython==2.2.0
email-validator==1.1.3
Flask==2.0.2
Flask-Login==0.5.0
Flask-Migrate==3.1.0
flask-restx==0.5.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
frozenlist==1.3.0
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.11.1
importlib-resources==5.4.0
isodate==0.6.1
itsdangerous==2.0.1
Jinja2==3.0.3
jsonschema==4.4.0
Mako==1.1.6
MarkupSafe==2.0.1
msrest==0.6.21
multidict==6.0.2
oauthlib==3.2.0
Pillow==9.0.1
psycopg2==2.9.3
pycparser==2.21
pyrsistent==0.18.1
python-decouple==3.5
python-dotenv==0.19.2
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
SQLAlchemy==1.4.29
typing_extensions==4.1.1
urllib3==1.26.8
Werkzeug==2.0.3
WTForms==3.0.1
yarl==1.7.2
zipp==3.7.0
